Revert "PIM-5210: Use the enhanced Product Edit Form in the mass edit common attributes action"
