--- features/Context/Page/Base/ProductEditForm.php
@@ -1,562 +1,0 @@
-<?php
-
-namespace Context\Page\Base;
-
-use Behat\Mink\Element\Element;
-use Behat\Mink\Element\NodeElement;
-use Behat\Mink\Exception\ElementNotFoundException;
-
-/**
- * Product Edit Form
- *
- * @author    Marie Bochu <marie.bochu@akeneo.com>
- * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class ProductEditForm extends Form
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function __construct($session, $pageFactory, $parameters = [])
-    {
-        parent::__construct($session, $pageFactory, $parameters);
-
-        $this->elements = array_merge(
-            $this->elements,
-            [
-                'Locales dropdown' => ['css' => '.attribute-edit-actions .locale-switcher'],
-                'Channel dropdown' => ['css' => '.attribute-edit-actions .scope-switcher'],
-            ]
-        );
-    }
-
-    /**
-     * @param string $attribute
-     * @param string $group
-     *
-     * @return NodeElement|null
-     */
-    public function findAvailableAttributeInGroup($attribute, $group)
-    {
-        $this->spin(function () {
-            return $this->find('css', $this->elements['Available attributes button']['css']);
-        }, 20, sprintf('Cannot find element "%s"', $this->elements['Available attributes button']['css']));
-
-        $list = $this->getElement('Available attributes list');
-        if (!$list->isVisible()) {
-            $this->openAvailableAttributesMenu();
-        }
-
-        $options = $this->spin(function () use ($list) {
-            return $list->findAll('css', 'li');
-        }, 20, 'No attributes found in available attributes list');
-
-        $groupedAttributes = [];
-        $currentOptgroup   = '';
-        foreach ($options as $option) {
-            if ($option->hasClass('ui-multiselect-optgroup-label')) {
-                $currentOptgroup = strtolower($option->getText());
-            } else {
-                $groupedAttributes[$currentOptgroup][$option->getText()] = $option;
-            }
-        }
-
-        $group = strtolower($group);
-        if (isset($groupedAttributes[$group]) && isset($groupedAttributes[$group][$attribute])) {
-            return $groupedAttributes[$group][$attribute];
-        }
-
-        return null;
-    }
-
-    /**
-     * This method allows to fill a field by passing the label
-     *
-     * @param string  $label
-     * @param string  $value
-     * @param Element $element
-     */
-    public function fillField($label, $value, Element $element = null)
-    {
-        $isLabel = false;
-
-        try {
-            $fieldContainer = $this->findFieldContainer($label);
-        } catch (ElementNotFoundException $e) {
-            $isLabel        = true;
-            $fieldContainer = $this->extractLabelElement($label, $element);
-        }
-
-        $fieldType = $this->getFieldType($fieldContainer, $isLabel);
-
-        switch ($fieldType) {
-            case 'text':
-            case 'date':
-            case 'number':
-                $this->fillTextField($fieldContainer, $value);
-                break;
-            case 'textArea':
-                $this->fillTextAreaField($fieldContainer, $value);
-                break;
-            case 'metric':
-                $this->fillMetricField($fieldContainer, $value);
-                break;
-            case 'multiSelect':
-                $this->fillMultiSelectField($fieldContainer, $value);
-                break;
-            case 'price':
-                $this->fillPriceField($fieldContainer, $value);
-                break;
-            case 'select':
-                $this->fillSelectField($fieldContainer, $value);
-                break;
-            default:
-                parent::fillField($fieldContainer->labelContent, $value);
-                break;
-        }
-    }
-
-    /**
-     * Find field container
-     *
-     * @param string $label
-     *
-     * @throws ElementNotFoundException
-     *
-     * @return NodeElement
-     */
-    public function findFieldContainer($label)
-    {
-        if (1 === preg_match('/ in (.{1,3})$/', $label)) {
-            // Price in EUR
-            $label = explode(' in ', $label)[0];
-        }
-
-        try {
-            $labelNode = $this->spin(function () use ($label) {
-                return $this->find('css', sprintf('.field-container header label:contains("%s")', $label));
-            }, 10);
-        } catch (\Exception $e) {
-            throw new ElementNotFoundException($this->getSession());
-        }
-
-        $container = $this->spin(function () use ($labelNode) {
-            return $labelNode->getParent()->getParent()->getParent();
-        });
-
-        $container->name = $label;
-
-        return $container;
-    }
-
-    /**
-     * @param string $label
-     * @param bool   $copy
-     *
-     * @throws ElementNotFoundException
-     *
-     * @return NodeElement
-     */
-    public function findField($label, $copy = false)
-    {
-        if (1 === preg_match('/ in (.{1,3})$/', $label)) {
-            // Price in EUR
-            list($label, $currency) = explode(' in ', $label);
-            $fieldContainer = $this->findFieldContainer($label);
-
-            return $this->findCompoundField($fieldContainer, $currency);
-        }
-
-        $subContainer = $this->spin(function () use ($label, $copy) {
-            return $this->findFieldContainer($label)
-                ->find('css', $copy ? '.copy-container .form-field' : '.form-field');
-        });
-
-        $field = $this->spin(function () use ($subContainer) {
-            return $subContainer->find('css', '.field-input input, .field-input textarea');
-        }, 10);
-
-        return $field;
-    }
-
-    /**
-     * Fills a textarea field element with $value
-     *
-     * @param NodeElement $fieldContainer
-     * @param string      $value
-     */
-    protected function fillTextAreaField(NodeElement $fieldContainer, $value)
-    {
-        $this->spin(function () use ($value, $fieldContainer) {
-            $field = $fieldContainer->find('css', 'div.field-input > textarea');
-
-            if (!$field || !$field->isVisible()) {
-                // the textarea can be hidden (display=none) when using WYSIWYG
-                $field = $fieldContainer->find('css', 'div.note-editor > .note-editable');
-            }
-
-            $field->setValue($value);
-
-            return ($field->getValue() === $value || $field->getHtml() === $value);
-        });
-
-        $this->getSession()->executeScript('$(\'.field-input textarea\').trigger(\'change\');');
-    }
-
-    /**
-     * Fills a simple select2 field with $value
-     *
-     * @param NodeElement $fieldContainer
-     * @param string      $value
-     *
-     * @throws ExpectationException
-     */
-    protected function fillSelectField(NodeElement $fieldContainer, $value)
-    {
-        if ('' === $value || null === $value) {
-            $emptyLink = $this->spin(function () use ($fieldContainer) {
-                return $fieldContainer->find('css', '.select2-search-choice-close');
-            });
-
-            $emptyLink->click();
-
-            $this->getSession()->executeScript(
-                '$(\'.field-input input[type="hidden"].select-field\').trigger(\'change\');'
-            );
-
-            return;
-        }
-
-        $link = $this->spin(function () use ($fieldContainer) {
-            return $fieldContainer->find('css', 'a.select2-choice');
-        }, 20, sprintf('Could not find select2 widget inside %s', $fieldContainer->getParent()->getHtml()));
-
-
-        $link->click();
-
-        $item = $this->spin(function () use ($link, $value) {
-            return $this->find('css', sprintf('.select2-results li:contains("%s")', $value));
-        });
-
-        $item->click();
-
-        $this->getSession()->executeScript(
-            '$(\'.field-input input[type="hidden"].select-field\').trigger(\'change\');'
-        );
-
-        return;
-    }
-
-    /**
-     * Fills a metric field with $value
-     *
-     * @param NodeElement $fieldContainer
-     * @param string      $value
-     *
-     * @throws \InvalidArgumentException
-     */
-    protected function fillMetricField(NodeElement $fieldContainer, $value)
-    {
-        if (false !== strpos($value, ' ')) {
-            list($text, $select) = explode(' ', $value);
-        } else {
-            $text   = $value;
-            $select = null;
-        }
-
-        $field = $fieldContainer->find('css', '.field-input');
-        if (null !== $select) {
-            if (null !== $link = $field->find('css', 'a.select2-choice')) {
-                $link->click();
-
-                $item = $this->spin(function () use ($select) {
-                    return $this->find('css', sprintf('#select2-drop li:contains("%s")', $select));
-                }, 5);
-            }
-
-            if (!$item) {
-                throw new \InvalidArgumentException(
-                    sprintf('Could not find select2 widget inside %s', $field->getParent()->getHtml())
-                );
-            }
-
-            $item->click();
-        }
-
-        $this->fillTextField($fieldContainer, $text);
-    }
-
-    /**
-     * Fills a select2 multi-select field with $values
-     *
-     *
-     * @param NodeElement $fieldContainer
-     * @param string      $values
-     *
-     * @throws \InvalidArgumentException
-     */
-    protected function fillMultiSelectField(NodeElement $fieldContainer, $values)
-    {
-        $field = $fieldContainer->find('css', '.form-field');
-
-        // clear multi select first
-        $fieldClasses = $field->getAttribute('class');
-        if (preg_match('/akeneo-multi-select(-reference-data)?-field/', $fieldClasses, $matches)) {
-            $select2Selector = sprintf('.%s div.field-input > input', $matches[0]);
-            $script          = sprintf('$("%s").select2("val", "");$("%1$s").trigger("change");', $select2Selector);
-            $this->getSession()->executeScript($script);
-        }
-
-        $link = $fieldContainer->find('css', 'ul.select2-choices');
-        if (null === $link) {
-            throw new \InvalidArgumentException(
-                sprintf('Could not find select2 widget inside %s', $fieldContainer->getParent()->getHtml())
-            );
-        }
-
-        foreach ($this->listToArray($values) as $value) {
-            $link->click();
-            $item = $this->spin(function () use ($value) {
-                return $this->find(
-                    'css',
-                    sprintf('.select2-result:not(.select2-selected) .select2-result-label:contains("%s")', $value)
-                );
-            }, 5);
-
-            // Select the value in the displayed dropdown
-            if (null !== $item) {
-                $item->click();
-            } else {
-                throw new \InvalidArgumentException(
-                    sprintf('Could not find select2 item with value %s inside %s', $value, $link->getHtml())
-                );
-            }
-        }
-
-        $this->getSession()->executeScript(
-            '$(\'.field-input input.select-field\').trigger(\'change\');'
-        );
-    }
-
-    /**
-     * Fills a compound field with $value, by passing the $label
-     *
-     * @param NodeElement $fieldContainer
-     * @param string      $value
-     *
-     * @throws ElementNotFoundException
-     */
-    protected function fillPriceField(NodeElement $fieldContainer, $value)
-    {
-        $amount   = null;
-        $currency = null;
-
-        if (false !== strpos($value, ' ')) {
-            list($amount, $currency) = explode(' ', $value);
-        }
-
-        if (!$currency) {
-            throw new \InvalidArgumentException(
-                sprintf(
-                    'The "%s" field is compound but the sub label was not provided',
-                    $fieldContainer->name
-                )
-            );
-        }
-
-        $field = $this->findCompoundField($fieldContainer, $currency);
-        $field->setValue($amount);
-
-        $this->getSession()->executeScript(
-            '$(\'.field-input input[type="text"]\').trigger(\'change\');'
-        );
-    }
-
-    /**
-     * Find a compound field
-     *
-     * @param NodeElement $fieldContainer
-     * @param             $currency
-     *
-     * @throws ElementNotFoundException
-     *
-     * @return NodeElement
-     */
-    protected function findCompoundField($fieldContainer, $currency)
-    {
-        $input = $fieldContainer->find('css', sprintf('input[data-currency=%s]', $currency));
-        if (!$input) {
-            throw new ElementNotFoundException(
-                $this->getSession(),
-                'compound field',
-                'id|name|label|value',
-                $fieldContainer->name
-            );
-        }
-
-        return $input;
-    }
-
-    /**
-     * Transform a list to array
-     *
-     * @param string $list
-     *
-     * @return array
-     */
-    public function listToArray($list)
-    {
-        if (empty($list)) {
-            return [];
-        }
-
-        return explode(', ', str_replace(' and ', ', ', $list));
-    }
-
-    /**
-     * Guesses the type of field identified by $label and returns it.
-     *
-     * Possible identified fields are :
-     * [date, metric, multiSelect, number, price, select, text, textArea]
-     *
-     * @param $fieldContainer
-     *
-     * @return string
-     */
-    protected function getFieldType($fieldContainer, $isLabel = false)
-    {
-        if (null === $fieldContainer || !$fieldContainer instanceof NodeElement) {
-            return null;
-        }
-
-        if ($isLabel) {
-            $formFieldWrapper = $fieldContainer->getParent()->getParent();
-        } else {
-            $formFieldWrapper = $fieldContainer->find('css', 'div.form-field');
-        }
-
-        if ($formFieldWrapper->hasClass('akeneo-datepicker-field')) {
-            return 'date';
-        } elseif ($formFieldWrapper->hasClass('akeneo-metric-field')) {
-            return 'metric';
-        } elseif ($formFieldWrapper->hasClass('akeneo-multi-select-field') ||
-                  $formFieldWrapper->hasClass('akeneo-multi-select-reference-data-field')
-        ) {
-            return 'multiSelect';
-        } elseif ($formFieldWrapper->hasClass('akeneo-number-field')) {
-            return 'number';
-        } elseif ($formFieldWrapper->hasClass('akeneo-price-collection-field')) {
-            return 'price';
-        } elseif ($formFieldWrapper->hasClass('akeneo-simple-select-field') ||
-                  $formFieldWrapper->hasClass('akeneo-simple-select-reference-data-field')
-        ) {
-            return 'select';
-        } elseif ($formFieldWrapper->hasClass('akeneo-text-field')) {
-            return 'text';
-        } elseif ($formFieldWrapper->hasClass('akeneo-textarea-field') ||
-                  $formFieldWrapper->hasClass('akeneo-wysiwyg-field')
-        ) {
-            return 'textArea';
-        } elseif ($formFieldWrapper->hasClass('akeneo-media-uploader-field')) {
-            return 'media';
-        } elseif ($formFieldWrapper->hasClass('akeneo-switch-field')) {
-            return 'switch';
-        } else {
-            return parent::getFieldType($fieldContainer);
-        }
-    }
-
-    /**
-     * Fills a text field element with $value, identified by its container or label.
-     *
-     * @param NodeElement $fieldContainerOrLabel
-     * @param string      $value
-     */
-    protected function fillTextField(NodeElement $fieldContainerOrLabel, $value)
-    {
-        $field = $fieldContainerOrLabel->find('css', 'div.field-input input');
-
-        // no field found, we're using a label
-        if (!$field) {
-            $field = $fieldContainerOrLabel->getParent()->getParent()->find('css', 'div.field-input input');
-        }
-
-        if (!$field) {
-            $field = $fieldContainerOrLabel->getParent()->find('css', 'div.controls input');
-        }
-
-        $field->setValue($value);
-        $this->getSession()->executeScript('$(\'.field-input input[type="text"]\').trigger(\'change\');');
-    }
-
-    /**
-     * @param string $localeCode
-     * @param bool   $copy
-     *
-     * @throws \Exception
-     */
-    public function switchLocale($localeCode, $copy = false)
-    {
-        $dropdown = $this->getElement(
-            $copy ? 'Copy locales dropdown' : 'Locales dropdown',
-            20,
-            'Could not find locale switcher'
-        );
-
-        $toggle = $this->spin(function () use ($dropdown) {
-            return $dropdown->find('css', '.dropdown-toggle');
-        });
-        $toggle->click();
-
-        $option = $this->spin(function () use ($dropdown, $localeCode) {
-            return $dropdown->find('css', sprintf('a[data-locale="%s"]', $localeCode));
-        }, 20, sprintf('Could not find locale "%s" in switcher', $localeCode));
-        $option->click();
-    }
-
-    /**
-     * @param string $scopeCode
-     * @param bool   $copy
-     *
-     * @throws \Exception
-     */
-    public function switchScope($scopeCode, $copy = false)
-    {
-        $dropdown = $this->getElement(
-            $copy ? 'Copy channel dropdown' : 'Channel dropdown',
-            20,
-            'Could not find scope switcher'
-        );
-
-        $toggle = $this->spin(function () use ($dropdown) {
-            return $dropdown->find('css', '.dropdown-toggle');
-        });
-        $toggle->click();
-
-        $option = $this->spin(function () use ($dropdown, $scopeCode) {
-            return $dropdown->find('css', sprintf('a[data-scope="%s"]', $scopeCode));
-        }, 20, sprintf('Could not find scope "%s" in switcher', $scopeCode));
-        $option->click();
-    }
-
-    /**
-     * Find a validation tooltip containing a text
-     *
-     * @param string $text
-     *
-     * @return null|Element
-     */
-    public function findValidationTooltip($text)
-    {
-        return $this->spin(function () use ($text) {
-            return $this->find(
-                'css',
-                sprintf(
-                    '.validation-errors span.error-message:contains("%s")',
-                    $text
-                )
-            );
-        });
-    }
-}
