+++ features/Context/Page/Batch/EditCommonAttributes.php
@@ -2,7 +2,7 @@
 
 namespace Context\Page\Batch;
 
-use Context\Page\Base\ProductEditForm;
+use Context\Page\Base\Wizard;
 
 /**
  * Edit common attributes page
@@ -11,50 +11,22 @@ use Context\Page\Base\ProductEditForm;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class EditCommonAttributes extends ProductEditForm
+class EditCommonAttributes extends Wizard
 {
-    protected $currentStep;
+    protected $elements = array(
+        'Available attributes button'     => array('css' => 'button:contains("Select attributes")'),
+        'Available attributes add button' => array('css' => '.pimmultiselect a:contains("Select")'),
+        'Available attributes form'       => array(
+            'css' => '#pim_enrich_mass_edit_choose_action_operation_displayedAttributes'
+        ),
+        'Locales dropdown' => array('css' => '#pim_enrich_mass_edit_choose_action_operation_locale'),
+    );
 
     /**
-     * {@inheritdoc}
+     * @param string $locale
      */
-    public function __construct($session, $pageFactory, $parameters = [])
+    public function switchLocale($locale)
     {
-        parent::__construct($session, $pageFactory, $parameters);
-
-        $this->elements = array_merge(
-            $this->elements,
-            [
-                'Available attributes button'     => ['css' => 'button:contains("Select attributes")'],
-                'Available attributes add button' => ['css' => '.pimmultiselect a:contains("Select")'],
-                'Available attributes form'       => [
-                    'css' => '#pim_enrich_mass_edit_choose_action_operation_displayedAttributes'
-                ],
-            ]
-        );
-    }
-
-    /**
-     * Go to the next step
-     *
-     * @return string
-     */
-    public function next()
-    {
-        $this->pressButton('Next');
-
-        return $this->currentStep;
-    }
-
-    /**
-     * Press the confirm button
-     *
-     * @return string
-     */
-    public function confirm()
-    {
-        $this->pressButton('Confirm');
-
-        return $this->currentStep;
+        $this->getElement('Locales dropdown')->selectOption($locale);
     }
 }
