+++ features/mass-action/update_common_attributes.feature
@@ -73,10 +73,9 @@ Feature: Edit common attributes of many products at once
   @jira https://akeneo.atlassian.net/browse/PIM-3281
   Scenario: Successfully update localized values on selected locale
     Given I add the "french" locale to the "mobile" channel
-    And I add the "french" locale to the "tablet" channel
     When I mass-edit products boots, sandals and sneakers
     And I choose the "Edit common attributes" operation
-    And I switch the locale to "fr_FR"
+    And I switch the locale to "French (France)"
     And I display the [name] attribute
     And I change the "[name]" to "chaussure"
     And I move on to the next step
@@ -92,11 +91,11 @@ Feature: Edit common attributes of many products at once
     And I set product "pump" family to "boots"
     And I mass-edit products boots and pump
     And I choose the "Edit common attributes" operation
-    And I switch the locale to "fr_FR"
+    And I switch the locale to "French (France)"
     And I display the [description] attribute
-    And I change the "[description]" to "Bar Fr"
-    And I switch the scope to "mobile"
-    And I change the "[description]" to "Foo Fr"
+    And I expand the "[description]" attribute
+    And fill in "pim_enrich_mass_edit_choose_action_operation_values_description_mobile_text" with "Foo Fr"
+    And fill in "pim_enrich_mass_edit_choose_action_operation_values_description_tablet_text" with "Bar Fr"
     And I move on to the next step
     And I wait for the "edit-common-attributes" mass-edit job to finish
     Then the french mobile description of "boots" should be "Foo Fr"
@@ -109,26 +108,25 @@ Feature: Edit common attributes of many products at once
     Given I mass-edit products boots and sandals
     And I choose the "Edit common attributes" operation
     And I display the Weight and Name attribute
-    And I visit the "Other" group
     And I change the "Weight" to "Edith"
     And I move on to the next step
-    Then I should see "There are errors in the attributes form"
-    And I should see the text "Product information"
-    And I should see the text "Weight"
-    And I should not see the text "Name"
-    When I am on the products page
+    Then I should see "Product information"
+    And I should see "Weight"
+    And I should see "Name"
+    When I am on the attributes page
+    And I am on the products page
     And I mass-edit products boots and sandals
     And I choose the "Edit common attributes" operation
-    Then I should not see the text "Product information"
-    And I should not see the text "Weight"
-    And I should not see the text "Name"
+    Then I should not see "Product information"
+    And I should not see "Weight"
+    And I should not see "Name"
 
   @jira https://akeneo.atlassian.net/browse/PIM-4777
   Scenario: Doing a mass edit of an attribute from a variant group does not override group value
     Given I mass-edit products highheels, blue_highheels and sandals
     And I choose the "Edit common attributes" operation
     And I display the Heel Height attribute
-    And I change the "Heel Height" to "3"
+    And fill in "Heel Height" with "3"
     And I move on to the next step
     And I wait for the "edit-common-attributes" mass-edit job to finish
     Then the metric "heel_height" of products highheels, blue_highheels should be "12"
