+++ features/mass-action/validate/validate_editing_common_numeric_attributes.feature
@@ -61,7 +61,9 @@ Feature: Validate editing common attributes of multiple products
     And I display the Length attribute
     And I change the Length to "foo"
     And I move on to the next step
-    Then I should see validation tooltip "This value should be a valid number."
+    # @TODO: fix this
+    And I display the Length attribute
+    Then I should see validation error "This value should be a valid number."
     Then the metric "Length" of products boots, sandals and sneakers should be ""
 
   Scenario: Successfully mass edit a number attribute
@@ -88,7 +90,7 @@ Feature: Validate editing common attributes of multiple products
     And I display the Number in stock attribute
     And I change the "Number in stock" to "-10"
     And I move on to the next step
-    Then I should see validation tooltip "This value should be 0 or more."
+    Then I should see validation error "This value should be 0 or more."
     And attribute number_in_stock of "boots" should be ""
     And attribute number_in_stock of "sandals" should be ""
     And attribute number_in_stock of "sneakers" should be ""
@@ -97,8 +99,8 @@ Feature: Validate editing common attributes of multiple products
     Given I mass-edit products boots and sneakers
     And I choose the "Edit common attributes" operation
     And I display the Price attribute
-    And I change the "Price" to "10 USD"
-    And I change the "Price" to "15 EUR"
+    And I change the "$ Price" to "10"
+    And I change the "€ Price" to "15"
     And I move on to the next step
     And I wait for the "edit-common-attributes" mass-edit job to finish
     Then the prices "Price" of products boots and sneakers should be:
@@ -119,9 +121,11 @@ Feature: Validate editing common attributes of multiple products
     And I mass-edit products boots, sandals and sneakers
     And I choose the "Edit common attributes" operation
     And I display the Price attribute
-    And I change the "Price" to "500 USD"
+    And I change the "$ Price" to "500"
     And I move on to the next step
-    Then I should see validation tooltip "This value should be 200 or less."
+    # @TODO: fix this
+    And I display the Price attribute
+    Then I should see validation error "This value should be 200 or less."
     Then the prices "Price" of products boots, sandals and sneakers should be:
       | amount | currency |
       |        | USD      |
