+++ spec/Pim/Bundle/EnrichBundle/Connector/Processor/MassEdit/Product/EditCommonAttributesProcessorSpec.php
@@ -4,7 +4,6 @@ namespace spec\Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Component\StorageUtils\Updater\ObjectUpdaterInterface;
 use Akeneo\Component\StorageUtils\Updater\PropertySetterInterface;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
@@ -26,16 +25,14 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         ValidatorInterface $validator,
         ProductMassActionRepositoryInterface $massActionRepository,
         AttributeRepositoryInterface $attributeRepository,
-        JobConfigurationRepositoryInterface $jobConfigurationRepo,
-        ObjectUpdaterInterface $productUpdater
+        JobConfigurationRepositoryInterface $jobConfigurationRepo
     ) {
         $this->beConstructedWith(
             $propertySetter,
             $validator,
             $massActionRepository,
             $attributeRepository,
-            $jobConfigurationRepo,
-            $productUpdater
+            $jobConfigurationRepo
         );
     }
 
@@ -51,7 +48,7 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
 
     function it_does_not_set_values_when_attribute_is_not_editable(
         $validator,
-        $productUpdater,
+        $propertySetter,
         AttributeInterface $attribute,
         AttributeRepositoryInterface $attributeRepository,
         ProductInterface $product,
@@ -71,24 +68,11 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         )->shouldBeCalled();
 
         $jobConfigurationRepo->findOneBy(['jobExecution' => $jobExecution])->willReturn($jobConfiguration);
-
-        $normalizedValues = addslashes(json_encode([
-            'categories' => [
-                [
-                    'scope' => null,
-                    'locale' => null,
-                    'data' => ['office', 'bedroom']
-                ]
-            ]
-        ]));
-
         $jobConfiguration->getConfiguration()->willReturn(
             json_encode(
                 [
                     'filters' => [],
-                    'actions' => [
-                        'normalized_values' => $normalizedValues
-                    ]
+                    'actions' => [['field' => 'categories', 'value' => ['office', 'bedroom'], 'options' => []]]
                 ]
             )
         );
@@ -96,16 +80,16 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         $violations = new ConstraintViolationList([]);
         $validator->validate($product)->willReturn($violations);
 
-        $attributeRepository->findOneByIdentifier('categories')->willReturn($attribute);
+        $attributeRepository->findOneBy(['code' => 'categories'])->willReturn($attribute);
         $product->isAttributeEditable($attribute)->willReturn(false);
-        $productUpdater->update($product, Argument::any())->shouldNotBeCalled();
+        $propertySetter->setData($product, 'categories', ['office', 'bedroom'], [])->shouldNotBeCalled();
 
         $this->process($product)->shouldReturn(null);
     }
 
     function it_sets_values_to_attributes(
         $validator,
-        $productUpdater,
+        $propertySetter,
         AttributeInterface $attribute,
         AttributeRepositoryInterface $attributeRepository,
         ProductInterface $product,
@@ -118,25 +102,11 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         $stepExecution->getJobExecution()->willReturn($jobExecution);
 
         $jobConfigurationRepo->findOneBy(['jobExecution' => $jobExecution])->willReturn($jobConfiguration);
-
-        $values = [
-            'categories' => [
-                [
-                    'scope' => null,
-                    'locale' => null,
-                    'data' => ['office', 'bedroom']
-                ]
-            ]
-        ];
-        $normalizedValues = addslashes(json_encode($values));
-
         $jobConfiguration->getConfiguration()->willReturn(
             json_encode(
                 [
                     'filters' => [],
-                    'actions' => [
-                        'normalized_values' => $normalizedValues
-                    ]
+                    'actions' => [['field' => 'categories', 'value' => ['office', 'bedroom'], 'options' => []]]
                 ]
             )
         );
@@ -144,16 +114,16 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         $violations = new ConstraintViolationList([]);
         $validator->validate($product)->willReturn($violations);
 
-        $attributeRepository->findOneByIdentifier('categories')->willReturn($attribute);
+        $attributeRepository->findOneBy(['code' => 'categories'])->willReturn($attribute);
         $product->isAttributeEditable($attribute)->willReturn(true);
-        $productUpdater->update($product, $values)->shouldBeCalled();
+        $propertySetter->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
 
         $this->process($product);
     }
 
     function it_sets_invalid_values_to_attributes(
         $validator,
-        $productUpdater,
+        $propertySetter,
         AttributeInterface $attribute,
         AttributeRepositoryInterface $attributeRepository,
         ProductInterface $product,
@@ -166,25 +136,11 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         $stepExecution->getJobExecution()->willReturn($jobExecution);
 
         $jobConfigurationRepo->findOneBy(['jobExecution' => $jobExecution])->willReturn($jobConfiguration);
-
-        $values = [
-            'categories' => [
-                [
-                    'scope' => null,
-                    'locale' => null,
-                    'data' => ['office', 'bedroom']
-                ]
-            ]
-        ];
-        $normalizedValues = addslashes(json_encode($values));
-
         $jobConfiguration->getConfiguration()->willReturn(
             json_encode(
                 [
                     'filters' => [],
-                    'actions' => [
-                        'normalized_values' => $normalizedValues
-                    ]
+                    'actions' => [['field' => 'categories', 'value' => ['office', 'bedroom'], 'options' => []]]
                 ]
             )
         );
@@ -194,10 +150,10 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         $violations = new ConstraintViolationList([$violation, $violation]);
         $validator->validate($product)->willReturn($violations);
 
-        $attributeRepository->findOneByIdentifier('categories')->willReturn($attribute);
+        $attributeRepository->findOneBy(['code' => 'categories'])->willReturn($attribute);
         $product->isAttributeEditable($attribute)->willReturn(true);
 
-        $productUpdater->update($product, $values)->shouldBeCalled();
+        $propertySetter->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
         $this->setStepExecution($stepExecution);
         $stepExecution->addWarning(Argument::cetera())->shouldBeCalled();
         $stepExecution->incrementSummaryInfo('skipped_products')->shouldBeCalled();
