+++ src/Pim/Bundle/EnrichBundle/Form/Type/MassEditAction/EditCommonAttributesType.php
@@ -3,9 +3,12 @@
 namespace Pim\Bundle\EnrichBundle\Form\Type\MassEditAction;
 
 use Pim\Bundle\CatalogBundle\Helper\LocaleHelper;
+use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\EnrichBundle\Form\View\ProductFormViewInterface;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\FormInterface;
+use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
@@ -58,7 +61,59 @@ class EditCommonAttributesType extends AbstractType
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder->add('values', 'hidden');
+        $builder
+            ->add(
+                'values',
+                'pim_enrich_localized_collection',
+                [
+                    'type'               => 'pim_product_value',
+                    'allow_add'          => false,
+                    'allow_delete'       => true,
+                    'by_reference'       => false,
+                    'cascade_validation' => true,
+                    'currentLocale'      => $options['current_locale']
+                ]
+            )
+            ->add(
+                'locale',
+                'entity',
+                [
+                    'choices' => $options['locales'],
+                    'class'   => $this->localeClassName,
+                    'select2' => true,
+                    'attr'    => [
+                        'class' => 'operation-param',
+                    ]
+                ]
+            )
+            ->add(
+                'displayedAttributes',
+                'entity',
+                [
+                    'class'        => $this->attributeClass,
+                    'choices'      => $options['all_attributes'],
+                    'required'     => false,
+                    'multiple'     => true,
+                    'expanded'     => false,
+                    'group_by'     => 'group.label',
+                    'choice_value' => function (AttributeInterface $attribute) {
+                        // Cast id to string to be compatible with ChoiceView
+                        return (string) $attribute->getId();
+                    }
+                ]
+            );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function finishView(FormView $view, FormInterface $form, array $options)
+    {
+        foreach ($view['locale']->vars['choices'] as $choice) {
+            $choice->label = $this->localeHelper->getLocaleLabel($choice->label);
+        }
+
+        $view->vars['groups'] = $this->productFormView->getView();
     }
 
     /**
@@ -66,7 +121,14 @@ class EditCommonAttributesType extends AbstractType
      */
     public function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setDefaults(['data_class' => $this->dataClass]);
+        $resolver->setDefaults(
+            [
+                'data_class'     => $this->dataClass,
+                'locales'        => [],
+                'all_attributes' => [],
+                'current_locale' => null
+            ]
+        );
     }
 
     /**
