+++ src/Pim/Bundle/EnrichBundle/MassEditAction/Operation/EditCommonAttributes.php
@@ -3,21 +3,18 @@
 namespace Pim\Bundle\EnrichBundle\MassEditAction\Operation;
 
 use Akeneo\Component\FileStorage\File\FileStorerInterface;
-use Akeneo\Component\StorageUtils\Updater\ObjectUpdaterInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Pim\Bundle\CatalogBundle\Builder\ProductBuilderInterface;
 use Pim\Bundle\CatalogBundle\Context\CatalogContext;
 use Pim\Bundle\CatalogBundle\Manager\ProductMassActionManager;
+use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\LocaleInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Pim\Bundle\UserBundle\Context\UserContext;
-use Symfony\Component\Filesystem\Filesystem;
+use Pim\Component\Catalog\FileStorage;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
-use Symfony\Component\Validator\Constraints\IsTrue;
-use Symfony\Component\Validator\Mapping\ClassMetadata;
-use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
  * Edit common attributes of given products
@@ -28,7 +25,7 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
  */
 class EditCommonAttributes extends AbstractMassEditOperation
 {
-    /** @var string */
+    /** @var ArrayCollection|ProductValueInterface[] */
     protected $values;
 
     /** @var ArrayCollection */
@@ -61,33 +58,14 @@ class EditCommonAttributes extends AbstractMassEditOperation
     /** @var ProductMassActionManager */
     protected $massActionManager;
 
-    /** @var string */
-    protected $errors;
-
-    /** @var ObjectUpdaterInterface */
-    protected $productUpdater;
-
-    /** @var ValidatorInterface */
-    protected $productValidator;
-
-    /** @var NormalizerInterface */
-    protected $internalNormalizer;
-
-    /** @var string */
-    protected $tmpStorageDir;
-
     /**
      * @param ProductBuilderInterface      $productBuilder
      * @param UserContext                  $userContext
      * @param CatalogContext               $catalogContext
      * @param AttributeRepositoryInterface $attributeRepository
      * @param NormalizerInterface          $normalizer
-     * @param FileStorerInterface          $fileStorer
+     * @param FileStorerInterface       $fileStorer
      * @param ProductMassActionManager     $massActionManager
-     * @param ObjectUpdaterInterface       $productUpdater
-     * @param ValidatorInterface           $productValidator
-     * @param NormalizerInterface          $internalNormalizer
-     * @param string                       $tmpStorageDir
      */
     public function __construct(
         ProductBuilderInterface $productBuilder,
@@ -96,33 +74,25 @@ class EditCommonAttributes extends AbstractMassEditOperation
         AttributeRepositoryInterface $attributeRepository,
         NormalizerInterface $normalizer,
         FileStorerInterface $fileStorer,
-        ProductMassActionManager $massActionManager,
-        ObjectUpdaterInterface $productUpdater,
-        ValidatorInterface $productValidator,
-        NormalizerInterface $internalNormalizer,
-        $tmpStorageDir = null
+        ProductMassActionManager $massActionManager
     ) {
         $this->productBuilder      = $productBuilder;
         $this->userContext         = $userContext;
         $this->catalogContext      = $catalogContext;
         $this->displayedAttributes = new ArrayCollection();
-        $this->values              = '';
+        $this->values              = new ArrayCollection();
         $this->normalizer          = $normalizer;
         $this->attributeRepository = $attributeRepository;
-        $this->fileStorer          = $fileStorer;
+        $this->fileStorer       = $fileStorer;
         $this->massActionManager   = $massActionManager;
-        $this->productUpdater      = $productUpdater;
-        $this->productValidator    = $productValidator;
-        $this->tmpStorageDir       = $tmpStorageDir;
-        $this->internalNormalizer  = $internalNormalizer;
     }
 
     /**
-     * @param string $values
+     * @param Collection $values
      *
-     * @return string
+     * @return EditCommonAttributes
      */
-    public function setValues($values)
+    public function setValues(Collection $values)
     {
         $this->values = $values;
 
@@ -130,7 +100,7 @@ class EditCommonAttributes extends AbstractMassEditOperation
     }
 
     /**
-     * @return string
+     * @return Collection
      */
     public function getValues()
     {
@@ -141,8 +111,6 @@ class EditCommonAttributes extends AbstractMassEditOperation
      * @param LocaleInterface $locale
      *
      * @return EditCommonAttributes
-     *
-     * @deprecated Will be removed in 1.6
      */
     public function setLocale(LocaleInterface $locale)
     {
@@ -153,8 +121,6 @@ class EditCommonAttributes extends AbstractMassEditOperation
 
     /**
      * @return LocaleInterface
-     *
-     * @deprecated Will be removed in 1.6
      */
     public function getLocale()
     {
@@ -169,8 +135,6 @@ class EditCommonAttributes extends AbstractMassEditOperation
      * @param Collection $displayedAttributes
      *
      * @return EditCommonAttributes
-     *
-     * @deprecated Will be removed in 1.6
      */
     public function setDisplayedAttributes(Collection $displayedAttributes)
     {
@@ -181,8 +145,6 @@ class EditCommonAttributes extends AbstractMassEditOperation
 
     /**
      * @return Collection
-     *
-     * @deprecated Will be removed in 1.6
      */
     public function getDisplayedAttributes()
     {
@@ -202,7 +164,11 @@ class EditCommonAttributes extends AbstractMassEditOperation
      */
     public function getFormOptions()
     {
-        return [];
+        return [
+            'locales'        => $this->userContext->getUserLocales(),
+            'all_attributes' => $this->getAllAttributes(),
+            'current_locale' => $this->getLocale()->getCode()
+        ];
     }
 
     /**
@@ -210,7 +176,17 @@ class EditCommonAttributes extends AbstractMassEditOperation
      */
     public function initialize()
     {
-        $this->values = '';
+        $locale = $this->getLocale()->getCode();
+        $this->catalogContext->setLocaleCode($locale);
+        $this->allAttributes = null;
+        $this->values = new ArrayCollection();
+
+        $allAttributes = $this->getAllAttributes();
+
+        $this->values = new ArrayCollection();
+        foreach ($allAttributes as $attribute) {
+            $this->addValues($attribute, $this->getLocale());
+        }
     }
 
     /**
@@ -221,32 +197,20 @@ class EditCommonAttributes extends AbstractMassEditOperation
      */
     public function finalize()
     {
-        $data = json_decode($this->values, true);
-        $filesystem = new Filesystem();
+        foreach ($this->getValues() as $productValue) {
+            $media = $productValue->getMedia();
 
-        foreach ($data as $attributeCode => $attributeValues) {
-            foreach ($attributeValues as $index => $value) {
-                if (isset($value['data']['filePath']) && '' !== $value['data']['filePath']) {
-                    $uploadedFile = new \SplFileInfo($value['data']['filePath']);
-                    $tmpStorageDir = (null !== $this->tmpStorageDir) ? $this->tmpStorageDir : sys_get_temp_dir() . '/pim/file_storage';
-                    $newPath = $tmpStorageDir . DIRECTORY_SEPARATOR . $uploadedFile->getFilename();
-
-                    $filesystem->rename($uploadedFile->getPathname(), $newPath);
-
-                    $data[$attributeCode][$index]['data']['filePath'] = $newPath;
-                }
+            if (null !== $media && null !== $media->getUploadedFile()) {
+                $file = $this->fileStorer->store($media->getUploadedFile(), FileStorage::CATALOG_STORAGE_ALIAS, true);
+                $productValue->setMedia($file);
             }
         }
-
-        $this->values = json_encode($data);
     }
 
     /**
      * Initializes self::allAttributes with values from the repository
      *
      * @return array
-     *
-     * @deprecated Will be removed in 1.6
      */
     public function getAllAttributes()
     {
@@ -271,6 +235,37 @@ class EditCommonAttributes extends AbstractMassEditOperation
     }
 
     /**
+     * Add all the values required by the given attribute
+     *
+     * @param AttributeInterface $attribute
+     * @param LocaleInterface    $locale
+     */
+    protected function addValues(AttributeInterface $attribute, $locale)
+    {
+        if ($attribute->isScopable()) {
+            foreach ($locale->getChannels() as $channel) {
+                $key = $attribute->getCode() . '_' . $channel->getCode();
+                $value = $this->productBuilder->createProductValue(
+                    $attribute,
+                    $locale->getCode(),
+                    $channel->getCode()
+                );
+
+                $this->productBuilder->addMissingPrices($value);
+                $this->values[$key] = $value;
+            }
+        } else {
+            $value = $this->productBuilder->createProductValue(
+                $attribute,
+                $locale->getCode()
+            );
+
+            $this->productBuilder->addMissingPrices($value);
+            $this->values[$attribute->getCode()] = $value;
+        }
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function getOperationAlias()
@@ -283,9 +278,19 @@ class EditCommonAttributes extends AbstractMassEditOperation
      */
     public function getActions()
     {
-        $actions = [
-            'normalized_values' => $this->getValues()
-        ];
+        $actions = [];
+
+        foreach ($this->values as $value) {
+            $rawData = $this->normalizer->normalize($value->getData(), 'json', ['entity' => 'product']);
+            // if the value is localizable, let's use the locale the user has chosen in the form
+            $locale = null !== $value->getLocale() ? $this->getLocale()->getCode() : null;
+
+            $actions[] = [
+                'field'   => $value->getAttribute()->getCode(),
+                'value'   => $rawData,
+                'options' => ['locale' => $locale, 'scope' => $value->getScope()]
+            ];
+        }
 
         return $actions;
     }
@@ -308,50 +313,5 @@ class EditCommonAttributes extends AbstractMassEditOperation
     public function getItemsName()
     {
         return 'product';
-    }
-
-    /**
-     * Add constraint on product values integrity.
-     *
-     * It registers constraint assertion that "hasValidValues" must return true on
-     * mass edit form submission.
-     *
-     * @param ClassMetadata $metadata
-     */
-    public static function loadValidatorMetadata(ClassMetadata $metadata)
-    {
-        $metadata->addGetterConstraint('validValues', new IsTrue([
-            'message' => 'mass_edit.edit_common_attributes.invalid_values'
-        ]));
-    }
-
-    /**
-     * Apply current values to a fake product and test its integrity with the product validator.
-     * If violations are raised, values are not valid.
-     *
-     * Errors are stored in json format to be useable by the Product Edit Form.
-     *
-     * @return bool
-     */
-    public function hasValidValues()
-    {
-        $data = json_decode($this->values, true);
-
-        $product = $this->productBuilder->createProduct('0');
-        $this->productUpdater->update($product, $data);
-        $violations = $this->productValidator->validate($product);
-
-        $errors = ['values' => $this->internalNormalizer->normalize($violations, 'internal_api')];
-        $this->errors = json_encode($errors);
-
-        return 0 === $violations->count();
-    }
-
-    /**
-     * @return string
-     */
-    public function getErrors()
-    {
-        return $this->errors;
     }
 }
