+++ src/Pim/Bundle/EnrichBundle/Resources/config/assets.yml
@@ -1,0 +1,3 @@
+css:
+    mass_edit:
+        - bundles/pimenrich/css/mass-edit.less
