+++ src/Pim/Bundle/EnrichBundle/Resources/config/connector/processors.yml
@@ -29,7 +29,6 @@ services:
             - '@pim_catalog.repository.product_mass_action'
             - '@pim_catalog.repository.attribute'
             - '@pim_connector.repository.job_configuration'
-            - '@pim_catalog.updater.product'
 
     pim_enrich.connector.processor.mass_edit.family.set_requirements:
         class: %pim_enrich.connector.processor.mass_edit.family.set_requirements.class%
