+++ src/Pim/Bundle/EnrichBundle/Resources/config/mass_actions.yml
@@ -47,10 +47,6 @@ services:
             - '@pim_serializer'
             - '@akeneo_file_storage.file_storage.file.file_storer'
             - '@pim_catalog.manager.product_mass_action'
-            - '@pim_catalog.updater.product'
-            - '@pim_catalog.validator.product'
-            - '@pim_internal_api_serializer'
-            - %tmp_storage_dir%
         tags:
             -
                 name: pim_enrich.mass_edit_action
