+++ src/Pim/Bundle/EnrichBundle/Resources/public/css/mass-edit.less
@@ -1,0 +1,3 @@
+.edit-common-attributes {
+  min-height: 400px;
+}
