+++ src/Pim/Bundle/EnrichBundle/Resources/public/js/product/form/attributes/add-attribute.js
@@ -26,21 +26,6 @@ define(
             tagName: 'div',
             className: 'add-attribute',
             template: _.template(template),
-            defaultOptions: {
-                title: _.__('pim_enrich.form.product.tab.attributes.btn.add_attributes'),
-                placeholder: _.__('pim_enrich.form.product.tab.attributes.info.search_attributes'),
-                buttonTitle: _.__('pim_enrich.form.product.tab.attributes.btn.add'),
-                emptyText: _.__('pim_enrich.form.product.tab.attributes.info.no_available_attributes'),
-                header: '',
-                height: 175,
-                minWidth: 225,
-                classes: 'pimmultiselect pim-add-attributes-multiselect',
-                position: {
-                    my: 'right top',
-                    at: 'right bottom',
-                    collision: 'none'
-                }
-            },
 
             /**
              * Render this extension
@@ -62,15 +47,27 @@ define(
             initializeSelectWidget: function () {
                 var $select = this.$('select');
 
-                var opts = this.defaultOptions;
-                opts.open = function () {
-                    var loadingMask = this.showLoadingMask();
-                    this.loadAttributesChoices()
-                        .always(function () {
-                            loadingMask.hide().$el.remove();
-                        });
-                }.bind(this);
-
+                var opts = {
+                    title: _.__('pim_enrich.form.product.tab.attributes.btn.add_attributes'),
+                    placeholder: _.__('pim_enrich.form.product.tab.attributes.info.search_attributes'),
+                    emptyText: _.__('pim_enrich.form.product.tab.attributes.info.no_available_attributes'),
+                    header: '',
+                    height: 175,
+                    minWidth: 225,
+                    classes: 'pimmultiselect pim-add-attributes-multiselect',
+                    position: {
+                        my: 'right top',
+                        at: 'right bottom',
+                        collision: 'none'
+                    },
+                    open: function () {
+                        var loadingMask = this.showLoadingMask();
+                        this.loadAttributesChoices()
+                            .always(function () {
+                                loadingMask.hide().$el.remove();
+                            });
+                    }.bind(this)
+                };
                 opts.selectedText     = opts.title;
                 opts.noneSelectedText = opts.title;
 
@@ -85,7 +82,7 @@ define(
                 var $footerContainer = $('<div>', { 'class': 'ui-multiselect-footer' });
                 var $saveButton = $('<a>', {
                     'class': 'btn btn-small',
-                    text: this.defaultOptions.buttonTitle
+                    text: _.__('pim_enrich.form.product.tab.attributes.btn.add')
                 }).on('click', function () {
                     $select.multiselect('close');
                     var values = $select.val();
