--- src/Pim/Bundle/EnrichBundle/Resources/translations/validators.en.yml
@@ -1,3 +1,0 @@
-mass_edit:
-    edit_common_attributes:
-        invalid_values: There are errors in the attributes form
