+++ src/Pim/Bundle/UIBundle/Resources/public/css/form.less
@@ -1125,24 +1125,11 @@ table#progress-table {
 .mass-edit-form {
   padding: 25px;
   background-color: lighten(@color-light_grey, 4%);
-  .product-edit-form {
-    min-height: 300px;
-  }
-  .product-edit-form .tabs-left > .sidebar > .nav-tabs li.header {
-    border-left: 0;
-  }
   hr {
     height: 3px;
     background-color: lighten(@color-light_grey, 2%);
     border-top: 1px solid lighten(@color-light_grey, 5%);
     border-bottom: 1px solid lighten(@color-light_grey, 4%);
-  }
-  .tab-content {
-    padding: 0;
-    border-left: solid 1px #ccc;
-  }
-  header {
-    display: none;
   }
 }
