+++ src/Pim/Bundle/UIBundle/Resources/public/css/pim.less
@@ -673,9 +673,11 @@ h3.tab-header {
     .display-flex();
 
     .other-actions {
-      .display-flex();
       .flex(1);
-      .justify-content(flex-end);
+      .flex-shrink(0);
+      .display-flex();
+      .align-items(flex-end);
+      .flex-direction(column);
     }
   }
